+++ sdk/identity/perf-tests/identity/test/index.spec.ts
@@ -1,10 +1,12 @@
 import { PerfStressProgram, selectPerfStressTest } from "@azure/test-utils-perfstress";
-import { DeviceCodeCredentialPersistenceTest } from "./DeviceCodeCredential/persistence.spec";
+import { ClientSecretCredentialPersistenceTest } from "./ClientSecretCredential/persistence.spec";
+import * as dotenv from "dotenv";
+dotenv.config();
 
 console.log("=== Starting the perfStress test ===");
 
 const perfStressProgram = new PerfStressProgram(
-  selectPerfStressTest([DeviceCodeCredentialPersistenceTest])
+  selectPerfStressTest([ClientSecretCredentialPersistenceTest])
 );
 
 perfStressProgram.run();