+++ sdk/identity/perf-tests/identity/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0-beta.1 (Unreleased)
 
-- Added a first performance test that does silent authentication with persistence enabled with `DeviceCodeCredental`.
+- Added a first performance test that does silent authentication with persistence enabled with `ClientSecretCredential`.